requirements.txt
openai
pymupdf
chromadb
tiktoken
pandas
streamlit
python-dotenv
tqdm
plotly
